Read the input image instead of using remote URL.
The returned JSON includes imageURL which we were using
to display the image as seen by the server. But we ran
into an installation where that URL was not accessible.
So this fix just reads the selected file for display.

Left a comment in the code for now to mention what we
used to do just in case.

Also updated npm dependencies.

Closes: #47